proof of concept and moving wait for condiiton
