+++ src/web/methods/index.ts
@@ -1,0 +1,6 @@
+import { click } from './interactions';
+
+const kakMethods = {
+  click,
+};
+export { kakMethods };
